refactor(certificate): consolidate endpoints into one function for Hobby limit
Vercel Hobby caps a deployment at 12 Serverless Functions; the 3 separate
api/certificate/* files pushed the total to 14 and failed the build. Merge
create + webhook + [code] into a single catch-all api/certificate/[...slug].ts
that routes by method + path segment (same public URLs, no frontend change).

Derive the path segment from req.url, not the filesystem catch-all param:
vercel.json's explicit /api/(.*) rewrite invokes the function without populating
req.query.slug. Verified on a preview deploy: create validates, :code 404s,
landing 200.
